LPS-159401 Move logic, no logical changes
