--- modules/util/source-formatter/src/main/java/com/liferay/source/formatter/check/PropertiesTestFileCheck.java
@@ -1,127 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.source.formatter.check;
-
-import com.liferay.petra.string.StringBundler;
-import com.liferay.petra.string.StringPool;
-import com.liferay.portal.kernel.util.NaturalOrderStringComparator;
-import com.liferay.portal.kernel.util.StringUtil;
-import com.liferay.portal.kernel.util.Validator;
-
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
-
-/**
- * @author Alan Huang
- */
-public class PropertiesTestFileCheck extends BaseFileCheck {
-
-	@Override
-	protected String doProcess(
-		String fileName, String absolutePath, String content) {
-
-		if (!fileName.endsWith("test.properties")) {
-			return content;
-		}
-
-		return _sortTestProperties(
-			fileName, content, StringPool.BLANK,
-			StringPool.POUND + StringPool.POUND);
-	}
-
-	private String _sortTestProperties(
-		String fileName, String content, String indent, String pounds) {
-
-		String indentWithPounds = indent + pounds;
-
-		CommentComparator comparator = new CommentComparator();
-
-		Pattern pattern = Pattern.compile(
-			StringBundler.concat(
-				"((?<=\\A|\n\n)", indentWithPounds, "\n", indentWithPounds,
-				"( .+)\n", indentWithPounds, "\n\n[\\s\\S]*?)(?=(\n\n",
-				indentWithPounds, "\n|\\Z))"));
-
-		Matcher matcher = pattern.matcher(content);
-
-		String previousProperties = null;
-		String previousPropertiesComment = null;
-		int previousPropertiesStartPosition = -1;
-
-		while (matcher.find()) {
-			String properties = matcher.group(1);
-			String propertiesComment = matcher.group(2);
-			int propertiesStartPosition = matcher.start();
-
-			if (pounds.length() == 2) {
-				String newProperties = _sortTestProperties(
-					fileName, properties, indent + StringPool.FOUR_SPACES,
-					StringPool.POUND);
-
-				if (!newProperties.equals(properties)) {
-					return StringUtil.replaceFirst(
-						content, properties, newProperties,
-						propertiesStartPosition);
-				}
-			}
-
-			if (Validator.isNull(previousProperties)) {
-				previousProperties = properties;
-				previousPropertiesComment = propertiesComment;
-				previousPropertiesStartPosition = propertiesStartPosition;
-
-				continue;
-			}
-
-			int value = comparator.compare(
-				previousPropertiesComment, propertiesComment);
-
-			if (value > 0) {
-				content = StringUtil.replaceFirst(
-					content, properties, previousProperties,
-					propertiesStartPosition);
-				content = StringUtil.replaceFirst(
-					content, previousProperties, properties,
-					previousPropertiesStartPosition);
-
-				return content;
-			}
-
-			previousProperties = properties;
-			previousPropertiesComment = propertiesComment;
-			previousPropertiesStartPosition = propertiesStartPosition;
-		}
-
-		return content;
-	}
-
-	private class CommentComparator extends NaturalOrderStringComparator {
-
-		@Override
-		public int compare(String comment1, String comment2) {
-			if (comment1.equals(" Default")) {
-				return -1;
-			}
-
-			if (comment2.equals(" Default")) {
-				return 1;
-			}
-
-			return super.compare(comment1, comment2);
-		}
-
-	}
-
-}
