+++ modules/util/source-formatter/src/main/resources/sourcechecks.xml
@@ -1344,10 +1344,6 @@
 			<category name="Styling" />
 			<description name="Applies rules to enforce consistency in code style." />
 		</check>
-		<check name="PropertiesTestFileCheck">
-			<category name="Styling" />
-			<description name="Performs several checks in `test.properties`." />
-		</check>
 		<check name="PropertiesWhitespaceCheck">
 			<category name="Styling" />
 			<description name="Finds missing and unnecessary whitespace in `.properties` files." />
